+++ plugins/pmsmart/src/test/java/org/itsallcode/whiterabbit/plugin/pmsmart/PMSmartExporterTest.java
@@ -1,6 +1,7 @@
 package org.itsallcode.whiterabbit.plugin.pmsmart;
 
 import static java.util.Arrays.asList;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -10,6 +11,7 @@ import java.time.LocalDate;
 import java.time.Month;
 import java.time.YearMonth;
 import java.util.Map;
+import java.util.Optional;
 
 import org.itsallcode.whiterabbit.api.PluginConfiguration;
 import org.itsallcode.whiterabbit.api.features.ProgressMonitor;
@@ -69,6 +71,26 @@ class PMSmartExporterTest
         runExport(day(date, DayType.WORK, activity(COST_CARRIER, Duration.ofHours(1), "project1")));
         verify(projectRowMock).enterComment(date, "project1");
         verify(projectRowMock).enterDuration(date, Duration.ofHours(1));
+    }
+
+    @Test
+    void exportActivityWithoutComments()
+    {
+        when(configMock.getOptionalValue("transfer.comments")).thenReturn(Optional.of("false"));
+
+        final LocalDate date = LocalDate.of(2021, Month.MAY, 3);
+        runExport(day(date, DayType.WORK, activity(COST_CARRIER, Duration.ofHours(1), "project1")));
+        verify(projectRowMock).enterDuration(date, Duration.ofHours(1));
+        verify(projectRowMock, never()).enterComment(any(), any());
+    }
+
+    @Test
+    void exportActivityWithCommentsSetToTrue()
+    {
+        final LocalDate date = LocalDate.of(2021, Month.MAY, 3);
+        runExport(day(date, DayType.WORK, activity(COST_CARRIER, Duration.ofHours(1), "project1")));
+        verify(projectRowMock).enterDuration(date, Duration.ofHours(1));
+        verify(projectRowMock).enterComment(date, "project1");
     }
 
     @Test